requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework==3.9.1
django-redis==4.10.0
mock==2.0.0
mockredispy==2.9.3
